spring-boot-data-jpa/src/main/java/com/base/service/EmployeeService.java
package com.base.service;

import com.base.model.Employee;
import com.base.repository.EmployeeRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;
import java.util.List;

@Service
@Transactional
public class EmployeeService {

    @Autowired
    EmployeeRepository employeeRepository;

    public List<Employee> findAll() {
        return employeeRepository.findAll();
    }

    public Employee findByName(String name) {
        return employeeRepository.findByName(name);
    }
}
